Compiles

+++ src/main/scala/edx/basic/blocks/TileTurntable.scala
@@ -36,7 +36,7 @@ class TileTurntable extends ResonantBlock(Material.piston) with TRotatable
 
   override def update()
   {
-    updateTurntableState(world, x, y, z)
+    updateTurntableState(world, x.toInt, y.toInt, z.toInt)
   }
 
   private def updateTurntableState(world: World, x: Int, y: Int, z: Int)
